+++ src/rAspCoreVueLauncher.Api/Data/AppDbContext.cs
@@ -1,8 +1,0 @@
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore;
-
-namespace rAspCoreVueLauncher.Api.Data;
-
-public sealed class AppDbContext(DbContextOptions<AppDbContext> options) : IdentityDbContext<AppUser>(options)
-{
-}