+++ src/rAspCoreVueLauncher.Api/Data/AppUser.cs
@@ -1,7 +1,0 @@
-using Microsoft.AspNetCore.Identity;
-
-namespace rAspCoreVueLauncher.Api.Data;
-
-public sealed class AppUser : IdentityUser
-{
-}